Majorly refactor the code base
This refactoring is intended to give the package hierarchy a more flat
structure. Some stylistic problems such as extra newlines at the end of
files were also fixed. More refactoring and cleanup will likely have to
be done in the coming commits, but this is the first and largest pass in
this effort.